requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
dynaconf==3.2.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
Flask-Session==0.8.0
redis==5.2.0
